refactor(frontend): separate task and common css
