+++ frontend/src/components/common.css
@@ -1,0 +1,12 @@
+h2, h3 {
+  color: #f3f4f6;
+}
+
+body {
+  color: #e5e7eb;
+}
+
+button {
+  cursor: pointer;
+}
+
