+++ frontend/src/features/tasks/components/EditTaskModal.css
@@ -78,3 +78,4 @@
   padding: 8px 12px;
   border-radius: 6px;
 }
+
