+++ frontend/src/features/tasks/components/EditTaskModal.tsx
@@ -1,4 +1,5 @@
-import "../../../components/EditTaskModal.css"
+import "../../../components/common.css"
+import "./EditTaskModal.css"
 import { useEffect, useState } from "react"
 import { updateTask } from "../../../lib/api"
 import type { Task } from "../../../types/task"
