+++ frontend/src/features/tasks/components/TaskList.css
@@ -1,15 +1,3 @@
-h2, h3 {
-  color: #f3f4f6;
-}
-
-body {
-  color: #e5e7eb;
-}
-
-button {
-  cursor: pointer;
-}
-
 .button-with-icon {
   display: flex;
   align-items: center;
@@ -144,3 +132,4 @@ button {
   margin-top: 8px;
   color: #f1f1f1;
 }
+
